test(launch): fold core beta tests into launch.test.ts
Todo 5's launch wiring tests landed in a new sibling file; the plan's F4 diff-surface allowlist names only launch.ts and the existing launch.test.ts. Pure structural merge: all 24 core beta cases move across unchanged, the electron mock is reconciled into one block, 45 tests total (21 + 24).
